+++ CHANGELOG.md
@@ -9,8 +9,27 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Added
 
+- Block: Add support for hidden for hidden form field in form block.
+
+### Fixed
+
+- Editor: Fix issue where users were unable to verify their accounts when logged in.
+
+## [0.7.0] - 2019-06-14
+
+### Added
+
 - Editor: Add version tag to the header.
+- Editor: Display the current Appsemble version in the toolbar.
+- Editor: Support for editing user profiles.
 - Helm: Add official Helm chart.
+- Server: Support database migrations.
+- Server: Support editing user profiles.
+
+### Changed
+
+- Editor: Move app templates to the server side.
+- Server: Move app templates to the server side.
 
 ### Fixed